@@ -1,14 +1,5 @@
-import { css } from '@emotion/react'
-import { TextField, TextFieldProps } from '@material-ui/core'
+import { TextField, TextFieldProps } from '@mui/material'
 import React from 'react'
-import { useClassName } from '~/utils/useClassName'
-
-export const toNumberOrNull = (str: string) =>
-  str !== '' && !isNaN(str as any) ? +str : null
-
-const numberInputStyle = css`
-  text-align: right;
-`
 
 type Props = TextFieldProps & {
   min?: number
@@ -16,13 +7,11 @@ type Props = TextFieldProps & {
 }
 
 const NumberField = ({ min, max, inputProps, ...props }: Props) => {
-  const numberInputClassName = useClassName(numberInputStyle)
-
   return (
     <TextField
       type="number"
       inputProps={{
-        className: numberInputClassName,
+        sx: { textAlign: 'right' },
         inputMode: 'numeric',
         pattern: '[0-9]*',
         min,
