@@ -1,8 +1,8 @@
 import { css } from '@emotion/react'
-import { Button } from '@material-ui/core'
+import { Button } from '@mui/material'
 import React, { useEffect, useRef, useState } from 'react'
 import ReactCrop, { Crop } from 'react-image-crop'
-import { CardType } from '~/modules/cards'
+import { SettingsCard } from '~/modules/cards'
 import getCroppedImage from '~/utils/getCroppedImage'
 
 const wrapperStyle = css`
@@ -23,7 +23,7 @@ const actionsStyle = css`
 `
 
 interface Props {
-  card: CardType
+  card: SettingsCard
   onRequestClose: () => void
   update: (src: string) => void
   crop: Crop
@@ -65,7 +65,7 @@ export default ({ card, onRequestClose, update, crop: orgCrop, setCrop: setOrgCr
           crop={crop}
           onChange={crop => setCrop(crop)}
           imageStyle={{ maxHeight: imageMaxHeight }}
-          onImageLoaded={el => imageEl.current = el}
+          onImageLoaded={el => { imageEl.current = el }}
         />
       </div>
       <div css={actionsStyle}>
