@@ -1,7 +1,6 @@
-import { createTheme, Theme as MuiTheme, ThemeProvider } from '@material-ui/core/styles'
+import { createTheme, Theme as MuiTheme, ThemeProvider } from '@mui/material'
 import { AppProps } from 'next/app'
 import React from 'react'
-import { RecoilRoot } from 'recoil'
 
 require('react-image-crop/dist/ReactCrop.css')
 
@@ -14,10 +13,8 @@ const theme = createTheme({})
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <RecoilRoot>
-      <ThemeProvider theme={theme}>
-        <Component {...pageProps} />
-      </ThemeProvider>
-    </RecoilRoot>
+    <ThemeProvider theme={theme}>
+      <Component {...pageProps} />
+    </ThemeProvider>
   )
 }
