@@ -6,21 +6,12 @@ module.exports = withMdx({
   basePath: process.env.BASE_PATH,
   trailingSlash: true,
   webpack(config) {
-    config.module.rules.push(
-      { test: /\.afm$/, use: 'raw-loader' },
-      { test: /node_modules\/(fontkit|linebreak)/, use: 'null-loader' },
-    )
-
+    // for jspdf.js
     config.resolve.alias = {
       ...config.resolve.alias,
-      fs: 'pdfkit/js/virtual-fs.js',
-    }
-
-    config.resolve.fallback= {
-      stream: require.resolve('stream-browserify'),
-      zlib: require.resolve('browserify-zlib'),
-      util: require.resolve('util/'),
-      assert: require.resolve('assert/'),
+      canvg: false,
+      html2canvas: false,
+      dompurify: false,
     }
 
     return config
