@@ -1,16 +1,16 @@
 import { css, Theme } from '@emotion/react'
-import React, { useEffect, useState } from 'react'
-import { useRecoilValue } from 'recoil'
+import React, { Dispatch, useEffect, useState } from 'react'
 import AddCard from './AddCard'
 import Card from './Card'
-import { cardsState, useCardsActions } from '~/modules/cards'
-import { defaultCountState } from '~/modules/settings'
+import DragOverlay from './DragOverlay'
+import { useAction } from '~/hooks/state'
+import { addCardsAction, removeCardAction, SettingsCard, updateCardCountAction, updateCardSrcAction } from '~/modules/cards'
 
 const cardsStyle = css`
   position: relative;
   @media (min-width: 600px) {
-      flex: 1;
-      overflow: auto;
+    flex: 1;
+    overflow: auto;
   }
 `
 
@@ -32,89 +32,83 @@ const emptyStyle = css`
 const footerStyle = (theme: Theme) => css`
   position: sticky;
   bottom: 0;
-  padding: ${theme.spacing(0, 1)};
+  padding: ${theme.spacing(1)};
+  background-color: #fff;
 `
 
-const overlayStyle = css`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  border: 4px dashed #666;
-  border-radius: 20px;
-  font-size: 1.5rem;
-  background-color: rgba(255, 255, 255, 0.7);
-  color: #666;
-  pointer-events: none;
-`
-
-const preventDefault = (e: Pick<Event, 'preventDefault'>) => {
-  e.preventDefault()
+interface Props {
+  cards: SettingsCard[]
+  cardInitCount: number
+  dispatch: Dispatch<any>
 }
 
-const Cards = () => {
-  const cards = useRecoilValue(cardsState)
-  const defaultCount = useRecoilValue(defaultCountState)
-  const { add, updateCount, updateSrc, remove } = useCardsActions()
+const Cards = ({ cards, cardInitCount, dispatch }: Props) => {
   const [isDraging, setDraging] = useState(false)
 
+  const addCards = useAction(addCardsAction, dispatch)
+  const updateCardCount = useAction(updateCardCountAction, dispatch)
+  const updateCardSrc = useAction(updateCardSrcAction, dispatch)
+  const removeCard = useAction(removeCardAction, dispatch)
+
   useEffect(() => {
-    const bodyListener = (e: Event) => {
-      setDraging(e.type === 'dragover')
+    const onDragOver = (e: DragEvent) => {
+      const data = e.dataTransfer
+
+      if (data && data.types.some(v => v === 'Files')) {
+        e.preventDefault()
+        setDraging(true)
+      }
+    }
+
+    const onDragLeave = () => {
+      setDraging(false)
+    }
+
+    const onDrop = (e: DragEvent) => {
+      e.preventDefault()
+
+      const fileList = Array.from(e.dataTransfer!.files)
+        .filter(file => file.type.startsWith('image/'))
+
+      addCards(fileList)
+      setDraging(false)
     }
 
     const body = document.body
 
-    body.addEventListener('dragover', bodyListener)
-    body.addEventListener('dragleave', bodyListener)
+    body.addEventListener('dragover', onDragOver)
+    body.addEventListener('dragleave', onDragLeave)
+    body.addEventListener('drop', onDrop)
 
     // destructor
     return () => {
-      body.removeEventListener('dragover', bodyListener)
-      body.removeEventListener('dragleave', bodyListener)
+      body.removeEventListener('dragover', onDragOver)
+      body.removeEventListener('dragleave', onDragLeave)
+      body.removeEventListener('drop', onDrop)
     }
-  }, [])
-
-  const onDrop = (e: React.DragEvent) => {
-    e.preventDefault()
-
-    const srcList = Array.from(e.dataTransfer!.files)
-      .filter(file => file.type.startsWith('image/'))
-      .map(file => URL.createObjectURL(file))
-
-    add(srcList)
-    setDraging(false)
-  }
+  }, [addCards])
 
   return (
-    <div css={cardsStyle} onDragOver={preventDefault} onDrop={onDrop}>
+    <div css={cardsStyle}>
       {!cards.length ? <p css={emptyStyle}>カード画像がありません</p> : (
         <ul css={listStyle}>
           {cards.map((card, index) =>
             <li key={card.id} css={itemStyle}>
               <Card
                 card={card}
-                defaultCount={defaultCount}
-                setCount={count => updateCount(index, count)}
-                setSrc={src => updateSrc(index, src)}
-                remove={() => remove(index)}
+                cardInitCount={cardInitCount}
+                setCount={count => updateCardCount(index, count)}
+                setSrc={src => updateCardSrc(index, src)}
+                remove={() => removeCard(index)}
               />
             </li>
           )}
         </ul>
       )}
       <div css={footerStyle}>
-        <AddCard add={add} fullWidth />
+        <AddCard add={addCards} fullWidth />
       </div>
-      {isDraging && (
-        <div css={[overlayStyle]} onDragOver={preventDefault}>
-          ここに画像をドロップ
-        </div>
-      )}
+      {isDraging && <DragOverlay />}
     </div>
   )
 }
